update tests
refactor

refactor buildroot
