+++ update-server/tests/openembedded/test_updater.py
@@ -28,6 +28,8 @@ def test_update(mock_root_fs_interface: MagicMock, mock_partition_manager: Magic
     mock_root_fs_interface.write_update.assert_called()
 
     updater.commit_update()
-
     mock_partition_manager.find_unused_partition.assert_called()
     mock_partition_manager.switch_partition.assert_called()
+
+    updater.write_update("/mmc/blk0p1", fake_callable(24))
+    mock_partition_manager.find_unused_partition.assert_called()
